requirements.txt
flask
flask-mysqldb
mysqldbclient
